+++ client/cache/cache.go
@@ -33,8 +33,9 @@ package cache
 
 import (
 	"context"
-	"fmt"
 	"time"
+
+	"github.com/beego/beego/v2/core/berror"
 )
 
 // Cache interface contains all behaviors for cache adapter.
@@ -55,12 +56,14 @@ type Cache interface {
 	// Set a cached value with key and expire time.
 	Put(ctx context.Context, key string, val interface{}, timeout time.Duration) error
 	// Delete cached value by key.
+	// Should not return error if key not found
 	Delete(ctx context.Context, key string) error
 	// Increment a cached int value by key, as a counter.
 	Incr(ctx context.Context, key string) error
 	// Decrement a cached int value by key, as a counter.
 	Decr(ctx context.Context, key string) error
 	// Check if a cached value exists or not.
+	// if key is expired, return (false, nil)
 	IsExist(ctx context.Context, key string) (bool, error)
 	// Clear all cache.
 	ClearAll(ctx context.Context) error
@@ -78,7 +81,7 @@ var adapters = make(map[string]Instance)
 // it panics.
 func Register(name string, adapter Instance) {
 	if adapter == nil {
-		panic("cache: Register adapter is nil")
+		panic(berror.Error(NilCacheAdapter, "cache: Register adapter is nil").Error())
 	}
 	if _, ok := adapters[name]; ok {
 		panic("cache: Register called twice for adapter " + name)
@@ -92,7 +95,7 @@ func Register(name string, adapter Instance) {
 func NewCache(adapterName, config string) (adapter Cache, err error) {
 	instanceFunc, ok := adapters[adapterName]
 	if !ok {
-		err = fmt.Errorf("cache: unknown adapter name %q (forgot to import?)", adapterName)
+		err = berror.Errorf(UnknownAdapter, "cache: unknown adapter name %s (forgot to import?)", adapterName)
 		return
 	}
 	adapter = instanceFunc()