+++ client/cache/calc_utils.go
@@ -1,46 +1,55 @@
 package cache
 
 import (
-	"fmt"
 	"math"
+
+	"github.com/beego/beego/v2/core/berror"
 )
+
+var (
+	ErrIncrementOverflow = berror.Error(IncrementOverflow, "this incr invocation will overflow.")
+	ErrDecrementOverflow = berror.Error(DecrementOverflow, "this decr invocation will overflow.")
+	ErrNotIntegerType = berror.Error(NotIntegerType, "item val is not (u)int (u)int32 (u)int64")
+)
+
+
 
 func incr(originVal interface{}) (interface{}, error) {
 	switch val := originVal.(type) {
 	case int:
 		tmp := val + 1
 		if val > 0 && tmp < 0 {
-			return nil, fmt.Errorf("increment would overflow")
+			return nil, ErrIncrementOverflow
 		}
 		return tmp, nil
 	case int32:
 		if val == math.MaxInt32 {
-			return nil, fmt.Errorf("increment would overflow")
+			return nil, ErrIncrementOverflow
 		}
 		return val + 1, nil
 	case int64:
 		if val == math.MaxInt64 {
-			return nil, fmt.Errorf("increment would overflow")
+			return nil, ErrIncrementOverflow
 		}
 		return val + 1, nil
 	case uint:
 		tmp := val + 1
 		if tmp < val {
-			return nil, fmt.Errorf("increment would overflow")
+			return nil, ErrIncrementOverflow
 		}
 		return tmp, nil
 	case uint32:
 		if val == math.MaxUint32 {
-			return nil, fmt.Errorf("increment would overflow")
+			return nil, ErrIncrementOverflow
 		}
 		return val + 1, nil
 	case uint64:
 		if val == math.MaxUint64 {
-			return nil, fmt.Errorf("increment would overflow")
+			return nil, ErrIncrementOverflow
 		}
 		return val + 1, nil
 	default:
-		return nil, fmt.Errorf("item val is not (u)int (u)int32 (u)int64")
+		return nil, ErrNotIntegerType
 	}
 }
 
@@ -49,35 +58,35 @@ func decr(originVal interface{}) (interface{}, error) {
 	case int:
 		tmp := val - 1
 		if val < 0 && tmp > 0 {
-			return nil, fmt.Errorf("decrement would overflow")
+			return nil, ErrDecrementOverflow
 		}
 		return tmp, nil
 	case int32:
 		if val == math.MinInt32 {
-			return nil, fmt.Errorf("decrement would overflow")
+			return nil, ErrDecrementOverflow
 		}
 		return val - 1, nil
 	case int64:
 		if val == math.MinInt64 {
-			return nil, fmt.Errorf("decrement would overflow")
+			return nil, ErrDecrementOverflow
 		}
 		return val - 1, nil
 	case uint:
 		if val == 0 {
-			return nil, fmt.Errorf("decrement would overflow")
+			return nil, ErrDecrementOverflow
 		}
 		return val - 1, nil
 	case uint32:
 		if val == 0 {
-			return nil, fmt.Errorf("increment would overflow")
+			return nil, ErrDecrementOverflow
 		}
 		return val - 1, nil
 	case uint64:
 		if val == 0 {
-			return nil, fmt.Errorf("increment would overflow")
+			return nil, ErrDecrementOverflow
 		}
 		return val - 1, nil
 	default:
-		return nil, fmt.Errorf("item val is not (u)int (u)int32 (u)int64")
+		return nil, ErrNotIntegerType
 	}
 }